copilot-prettyprinter: Create new library for pretty-printer. Refs #383.
The architecture of Copilot mixes backends with other libraries, which
makes the result not uniform. More specifically, the pretty-printer is
part of copilot-core, while the C99 backend and the interpreter are in
separate libraries.

The implementation does not match our desired architecture according to
our design documents, and it makes the code partly harder to maintain.

This commit introduces a new library, copilot-prettyprinter, that
publishes functions to pretty-print a Copilot Core expression. The
modules are taken from copilot-core and published under a different
namespace.