copilot-core: Deprecate Copilot.Core.PrettyPrint. Refs #383.
The architecture of Copilot mixes backends with other libraries, which
makes the result not uniform. More specifically, this library publishes
a pretty-printer under the Copilot.Core namespace, while the C99 backend
and the interpreter are in separate libraries and under separate
namespaces.

The implementation does not match our desired architecture according to
our design documents, and it makes the code partly harder to maintain.

This commit deprecates the pretty-printer included in copilot-core, and
directs users to the one published by the new library
copilot-prettyprinter.