copilot-c99: Remove struct declarations from `.c` file. Refs #373.
Copilot-generated C99 monitor implementation files import the
corresponding generated headers before defining the necessary C structs.
Part of those headers contain the declarations of the handlers, which
may take structs as arguments.

By using such forward declarations, the C compiler is not always able to
figure out the size of the arguments to those handler functions, and the
code cannot be compiled.

This commit removes the struct declarations from the `.c` file, so that
we can later move the declarations to an external header file and
include them before the external handlers are declared.