copilot-c99: Write type declarations in separate header file. Refs #373.
Copilot-generated C99 monitor implementation files import the
corresponding generated headers before defining the necessary C structs.
Part of those headers contain the declarations of the handlers, which
may take structs as arguments.

By using such forward declarations, the C compiler is not always able to
figure out the size of the arguments to those handler functions, and the
code cannot be compiled.

This commit adds an action to write the type declarations to a separate
header file.