copilot-c99: Include type declarations in generated `.c` file. Refs #373.
Copilot-generated C99 monitor implementation files import the
corresponding generated headers before defining the necessary C structs.
Part of those headers contain the declarations of the handlers, which
may take structs as arguments.

By using such forward declarations, the C compiler is not always able to
figure out the size of the arguments to those handler functions, and the
code cannot be compiled.

This commit includes the generated type declarations header file before
the header file that declares the handlers. This ensures that all the
types are declared before the functions that use them are declared.