copilot-core: Deprecate module Copilot.Core.Type.Equality. Refs #379.
The module Copilot.Core.Type.Equality provides definitions to reason
about type equality. That module is very similar to
base:Data.Type.Equality, and all uses of the former have been replaced
by uses of the latter. Consequently, the former is no longer used in
Copilot, and it is not required by Copilot's requirements.

This commit deprecates the module Copilot.Core.Type.Equality so that it
can be removed in a future version of Copilot.