copilot-core: Disable warning due to import of deprecated module. Refs #379.
The module Copilot.Core.Type.Equality has been deprecated, as it is no
longer needed due to the same functionality being available in base.

Our transition policy requires 2 versions to continue offering the
deprecated API until it can be properly removed. Due to an instance of a
class from Copilot.Core.Type.Equality being defined in
Copilot.Core.Type, and due to the former module now being deprecated,
the import of the module is causing compilation warnings.

This commit temporarily disables such warnings, adding a note to help
understand the reason for that adaptation and when it can be removed.