copilot-core: Add test for simpleType. Refs #393.
The function Copilot.Core.Type.simpleType previously returned SBool for the
simple type representation of an Int8 :: Typed Int8, when it should instead
return SInt8.

This error was not not caught by the tests because that specific aspect of that
function was not being tested.

This commit adds a regression test for simpleType that will catch that error,
as well as similar errors for simpleType. We do not consider the case of
structs or arrays due to the complexity of generating random values of those
types.