copilot-core: Remove test for module Copilot.Core.External. Refs #391.
The module Copilot.Core.External is not being used by any other part of
Copilot. It was deprecated in Copilot 3.10 and no messages have been received
requesting that it be kept in the implementation. As per our internal policy of
waiting 3 versions from deprecation until a public interface declaration can be
removed, this module can now be removed.

This commit removes all tests associated to that module, so that it can
be removed safely.