copilot-core: Remove module Copilot.Core.External. Refs #391.
The module Copilot.Core.External is not being used by any other part of
Copilot. It was deprecated in Copilot 3.10 and no messages have been received
requesting that it be kept in the implementation. As per our internal policy of
waiting 3 versions from deprecation until a public interface declaration can be
removed, this module can now be removed.

This commit removes the module completely. A notice to allow for deprecations
in Copilot.Core is also being removed, since it was added only to allow
re-exporting of Copilot.Core.External without warnings.