copilot-core: Remove dependency from cabal file. Refs #391.
The module Copilot.Core.External is not being used by any other part of
Copilot. It was deprecated in Copilot 3.10 and no messages have been received
requesting that it be kept in the implementation. As per our internal policy of
waiting 3 versions from deprecation until a public interface declaration can be
removed, this module can now be removed.

This commit removes the dependency on the package dlist, which was only used by
the now-removed Copilot.Core.External.