copilot-core: Remove tests involving Copilot.Core.Type.Show. Refs #392.
The module Copilot.Core.Type.Show is not being used outside of copilot-core. It
was deprecated in Copilot 3.10 and no messages have been received requesting
that it be kept in the implementation. As per our internal policy of waiting 3
versions from deprecation until a public interface declaration can be removed,
this module can now be completely hidden from the public interface.

This commit removes all tests associated to or using that module, so that it
can be removed safely. Tests related to the interpreter are removed as well,
since they are also present in the copilot-interpreter package.