copilot-c99: Use pointer to pass output array as argument to generators. Refs #386.
Copilot's C99 backend generates incorrect code. The functions that generate
values used to update variable buffers (which are ultimately used also to pass
arguments to trigger functions) may return pointers to stack-allocated arrays.
That memory may be used for something else after the call to such "generator"
functions, so it may be overwritten or freed by the time the value is used to
update the buffer. This may lead to buffers holding incorrect values, and also
to calls to trigger functions with incorrect arguments.

This commit changes generator functions for arrays so that they receive a
pointer to an array for the output argument. The alternative implementation
passes the responsibility of allocating the array to the caller.

The part of the generated code to update the values of buffers is adjusted
accordingly to provide pre-allocated arrays as arguments.

Two local functions in the CodeGen module (memcpy and tyElemName) are promoted
from the scope of the function mkstep to the top-level of the module, as they
are needed by a newly introduced function.