copilot-c99: Declare local array variables in generated guards as pointers. Refs #401.
Copilot's c99 backend is producing invalid code for variable initializers when
local is used. The code produced for temporary variables of type array
indicates their length in the type, but the value they are initialized to is a
compound literal, which is not a valid initialization expression.

This patch modifies the translation of a local or temporary variable of array
type to use a pointer for (the outermost layer of) arrays. The use of a pointer
allows us to use any expression to initialize the variable on the right-hand
side, making the current expressions, which use compound literals, valid.

Note that the word local is used in this context, both in the commit message
and in the code, with two different meanings: one is Copilot's local operator,
used to define "let expressions", and the other is the notion of a local
variable in C.