copilot: Re-structure README. Refs #415.
Copilot's README contains outdated information and is not adapted for
our current target audience. The instructions should be updated to
reflect up-to-date installation methods, be more approachable for
non-Haskellers, and describe the current software development processes
(e.g., how to file issues or make contributions, if possible).

This commit re-structures the README with the following changes:

- Styled header
- List of features
- TOC
- Updated installation instructions
- Examples (minor tweaks)
- Related projects
- Documentation (including papers)
- How to contribute
- Contributors and acknowledgements

The license is removed since it is available in each package.