copilot-core: Remove Copilot.Core.PrettyDot module. Refs #409.
The module Copilot.Core.PrettyDot is not being used by any other part of
Copilot. It was deprecated in Copilot 3.11 and no messages have been received
requesting that it be kept in the implementation. As per our internal policy of
waiting 3 versions from deprecation until a public interface declaration can be
removed, this module can now be removed.

This commit removes the module completely.