copilot-language: Remove dependency on copilot-prettyprinter. Refs #412.
The function Copilot.Language.prettyPrint is not being used by any other part
of Copilot. It was deprecated in Copilot 3.11 and no messages have been
received requesting that it be kept in the implementation. As per our internal
policy of waiting 3 versions from deprecation until a public interface
declaration can be removed, this function can be removed.

A prior commit removed the function. The library copilot-language no longer
imports any definitions from copilot-prettyprinter, so that library can now be
removed from the dependency list in the cabal file.