copilot-language: Remove Copilot.Language.prettyPrint. Refs #412.
The function Copilot.Language.prettyPrint is not being used by any other part of
Copilot. It was deprecated in Copilot 3.11 and no messages have been received
requesting that it be kept in the implementation. As per our internal policy of
waiting 3 versions from deprecation until a public interface declaration can be
removed, this function can now be removed.

This commit removes the function completely. The function prettyPrint, which
was imported from copilot-prettyprinter, is no longer needed.