copilot-core: Remove module space Copilot.Core.Interpret. Refs #410.
The modules Copilot.Core.Interpret and Copilot.Core.Interpret.Eval are
not being used by any other part of Copilot. They were deprecated in
Copilot 3.11 and their code moved to the copilot-interpreter library. As
per our internal policy of waiting 3 versions from deprecation until a
public interface declaration can be removed, this module can now be
removed.

This commit removes all modules in copilot-core specific to the
interpreter, and adjusts the Cabal file so that there are no references
to them.