copilot-core: Update reference to interpreter in docs. Refs #410.
All modules under the namespace Copilot.Core.Interpret have been removed
from copilot-core. The interpreter is now in the library
copilot-interpreter, under a different module name space.

This commit updates a reference in the documentation that pointed to
Copilot.Core.Interpret, so that it instead refers to the library where
that implementation now lives.

We cannot point to the new module name specifically because haddock
cannot resolve a link to a module we do not depend on (a module which,
depends on copilot-core itself). We instead choose to just name the
whole library, since it's dedicated to the interpreter and fits the
purpose intended in the comment.