copilot-theorem: Bump constraint on what4. Refs #423.
In order to keep Copilot effectively working in the current Haskell
ecosystem, as well as new versions of GHC, we need to extend the
versions of dependencies that Copilot can be installed with. For
example, what4 has seen a new release 0.14, but copilot-theorem needs
versions strictly lower than 0.14. We also do not support GHC 9.4 or
above.

This commit bumps te upper bounds on what so that it can be installed
with the 1.4 series.