copilot-core: Remove uses of Copilot.Core.Type.Equality. Refs #427.
The module Copilot.Core.Type.Equality is not being used by any other
part of Copilot. It's implementation was replaced by definitions
existing in the base library. The implementation in copilot-core was
deprecated in Copilot 3.12 and no messages have been received requesting
that it be kept in this library. As per our internal policy of waiting 3
versions from deprecation until a public interface declaration can be
removed, this module can now be removed.

This commit removes all uses of definitions from
Copilot.Core.Type.Equality, which were being kept in copilot-core to
facilitate a transition to the new interface based on Data.Type.Equality
while respecting our deprecation policy. Definitions equivalent to the
ones removed, but based on Data.Type.Equality, are still made available.