copilot-core: Update reference to point to copilot-prettyprinter. Refs #426.
The module Copilot.Core.PrettyPrint is not being used by any other part
of Copilot. It's implementation was moved to a dedicated library
copilot-prettyprint. The implementation in copilot-core was deprecated
in Copilot 3.12 and no messages have been received requesting that it be
kept in this library. As per our internal policy of waiting 3 versions
from deprecation until a public interface declaration can be removed,
this module can now be removed.

This commit updates a reference in a comment to point to the library
copilot-prettyprinter, instead of the module Copilot.Core.PrettyPrint.