copilot-c99: Add testing infrastructure. Refs #413.
As part of the requirements for Class D (NPR7150.2C), we need unit tests
for all of copilot.

This commit adds tests for copilot-c99, including the following cases:

- A valid specification is compiled correctly (without errors).
- A valid specification generates C code that can be compiled correctly.
- A valid specification generates C code that behaves as expected.

We test both normal streams (using plain types), as well as streams
carrying arrays.

Tests are enabled in the Travis CI infrastructure so that they run on
every pull request.

We do not test structs (yet) due to the general difficulty in generating
structs using QuickCheck, a limitation we found when building tests for
other parts of Copilot. That feature is deferred to future work.