copilot-language: Move Copilot.Language.Stream.Arg to Copilot.Language.Spec. Refs #446.
The type Copilot.Language.Stream.Arg is defined in the same module as a
Stream, but nothing in that module uses an Arg. Architecturally, that
type doesn't belong in that module, but, rather, should be listed in
Copilot.Language.Spec, where it is used in the definition of Trigger.

Consequently, Copilot.Language.Stream.Arg should be moved to
Copilot.Language.Spec.

As per our internal policy of waiting 3 versions from deprecation until
a public interface declaration can be hidden, this definition should be
deprecated in Copilot.Language.Stream.Arg and re-published from
Copilot.Language.Spec. However, Haskell offers no mechanism to do that.
If we deprecate it in Copilot.Language.Stream.Arg, we cannot make that
warning disappear when imported from Copilot.Language.Spec, even with
{-# OPTIONS_GHC -fno-warn-deprecations #-}.

This commit moves Copilot.Language.Stream.Arg to Copilot.Language.Spec.
All imports are adjusted accordingly.