copilot-libraries: Fix semantics of since in Copilot.Library.PTLTL. Refs #443.
The current implementation of `Copilot.Library.PTLTL.since` does not
conform to the standard semantics of since, as a temporal operator.
`since` is an existential claim; it must be true when there exists *any*
prior period in which s2 holds and after which s1 has been continuously
true. The current implementation only looks for the first time s2 is
true, which is non-standard.

This commit fixes the implementation and updates the comment to reflect
the correct semantics.