copilot-theorem: Relax version constraint on what4. Refs #461.
In order to keep Copilot effectively working in the current Haskell ecosystem,
we need to extend the versions of dependencies that Copilot can be installed
with. For example, what4 has seen a new release 1.5 and 1.5.1, but
copilot-theorem needs versions strictly lower than 1.5.

This commit relaxes the upper bounds on what4 so that it can be installed with
the 1.5 series.