copilot-core: Rename tylength to typeLength. Refs #457.
To comply with our NASA Class D requirements, our code must abide by a style
guide. More specifically, the function `tylength` is not lowerCamelCase, and
its name is not uniform (other functions in the same module spell out the word
type).

This commit introduces a function `typeLength` that replaces the old
`tylength`. In compliance with our policy of waiting 3 versions from
deprecation until a public interface declaration can be removed, the old
function is deprecated.