copilot-core: Rename typename to typeName. Refs #457.
To comply with our NASA Class D requirements, our code must abide by a style
guide. More specifically, the function `typename` is not lowerCamelCase, which
we require to be.

This commit introduces a function `typeName` that replaces the old `typename`.
In compliance with our policy of waiting 3 versions from deprecation until a
public interface declaration can be removed, the old function is deprecated.