copilot-core: Rename accessorname to accessorName. Refs #457.
To comply with our NASA Class D requirements, our code must abide by a style
guide. More specifically, the function `accessorname` is not lowerCamelCase,
which we require to be.

This commit introduces a function `accessorName` that replaces the old
`accessorname`.  In compliance with our policy of waiting 3 versions from
deprecation until a public interface declaration can be removed, the old
function is deprecated.