copilot-theorem: Introduce testing infrastructure for Copilot.Theorem.What4. Refs #474.
As part of the requirements for Class D (NPR7150.2C), we need unit tests
for all of copilot.

This commit introduces testing infrastructure for Copilot.Theorem.What4.
The module only defines some basic tests using Z3, but it does not
exploit all possible solvers supported by What4, or other connections
provided by other modules in copilot-theorem. We do so to limit the
scope (and thefore difficulty) of the change, and because old modules
that provided backends that we can now connect through What4 will likely
disappear from future versions of copilot-theorem.