copilot-c99: Pass output arrays as arguments to trigger argument functions. Refs #431.
Copilot's C99 backend generates incorrect code for trigger argument functions
that return constant arrays. Currently, the code will return a pointer to a
stack-allocated array. That memory may be used for something else after the
call to the trigger argument function, so it may be overwritten or freed by the
time the array is used. This may lead to the array holding incorrect values.

This is almost the same issue that was noticed in #386, except that #386 was
specifically about stream generator functions, whereas this issue is
specifically about trigger argument functions.

This commit changes array trigger argument functions so that they receive a
pointer to an array for the output argument. The part of the generated code
that applies the trigger argument functions has also been updated accordingly
such that when the argument has an array type, the output array is passed
directly as an argument (e.g., `f_arg0(temp_array)`) rather than being assigned
(e.g., `temp_array = f_arg0()`).

Because we pass output arrays directly to trigger argument functions, we no
longer have to have a special case in the code for declaring temporary
variables for output arrays as pointers (e.g., `int* temp_array;`). Instead, we
can declare these temporary variables directly as arrays (e.g., `int
temp_array[2];`). This makes the intent clearer and simplifies the
implementation in `copilot-c99`.