copilot-c99: Change return type of main generated for tests. Refs #468.
copilot-c99's tests sometimes fail because the return code of the programs
being generated and executed during tests is not zero. When such tests are
executed, the test interprets non-zero exit codes as failures. More
specifically, the return type of the main function included in the generated C
code is void, and compilers are not mandated to end the process with a zero
return code in such cases.

This commit modifies the type of the main function used during tests to make it
an int. C compilers will then make the exit code zero by default.