copilot-c99: Refactor variable update functionality. Refs #431.
The previous commit introduced functionality for handling array arguments in
trigger argument functions as a special case. The special-casing needed to
handle array arguments in trigger argument functions is nearly identical to the
special-casing needed in stream generator functions. This patch refactors the
code to share the special-casing in common between the two sorts of functions.