copilot-libraries: Introduce testing infrastructure for Copilot.Library. Refs #475.
As part of the requirements for Class D (NPR7150.2C), we need unit tests
for all of Copilot.

This commit introduces testing infrastructure for Copilot.Library. The
module only defines some basic tests specifically for the PTLTL module,
using both evaluation (interpretation) and an SMT-solver (via what4,
with Z3). The commit only seeks to introduce the infrastructure needed
to test other definitions provided by the library, and to demonstrate
how such infrastructure can be used.