copilot-c99: Adapt backend to work with language-c99-simple 0.3. Refs #472.
It has been determined that Copilot-generated code should comply with
the MISRA C coding standard.

The code generated by the C99 backend does not specify that guards and
generator functions are internal to the module, even though they are
only meant to be called by `step`. However, the lack of a storage class
implies that they are implicitly defined with external linkage,
violating several rules in MISRA (8.4 and 8.7). In order to fix that, we
need to specify that such functions are static, but the
language-c99-simple library only incorporated the ability to specify
storage information for function definitions in version 0.3.

This commit updates the cabal file of the C99 backend to use
language-c99-simple >= 0.3. It also updates all uses affected by the
change in API, but does not otherwise alter the behavior of the C99
backend.