copilot-c99: Make functions internal to the module static. Refs #472.
It has been determined that Copilot-generated code should comply with
the MISRA C coding standard.

The code generated by the C99 backend does not specify that guards and
generator functions are internal to the module, even though they are
only meant to be called by `step`. However, the lack of a storage class
implies that they are implicitly defined with external linkage,
violating several rules in MISRA (8.4 and 8.7).

This commit updates the guards and generators so that they are marked as
static and therefore internal to the module in which they are defined.
This addresses the aforementioned violations.