copilot-c99: Cast constants of type size_t used in math operations. Refs #472.
It has been determined that Copilot-generated code should comply with
the MISRA C coding standard.

The code generated by the C99 backend does not specify the types of all
constants used in specifications. Specifically, constants are being used
in the calculations needed to access and advance the ring buffers, but
the type of such constants, which is supposed to be `size_t`, is not
specified. This is a possible violation of MISRA C 2012 Rule 10.4, and
cppcheck reports it as so (although other tools do not report this
violation).

This commit updates the functions that access or update ring buffers to
specify the types used in constants referring to the index in the
buffer or its length, thus solving the existing violation of rule 10.4.