copilot: Update CI job to check for MISRA compliance with cppcheck. Refs #472.
It has been determined that Copilot-generated code should comply with
the MISRA C coding standard.

To ensure that future versions of Copilot keep complying with MISRA,
this commit modifies the CI job to check a sample file for compliance
using cppcheck.

Note that the version of cppcheck included with xenial, which is the
distribution used currently by Travis by default, does not include a
version of cppcheck modern enough to accept the flag --addon=misra.py.
Consequently, we update the dist to be focal.