copilot: Relax version constraint on optparse-applicative. Refs #488.
In order to keep Copilot effectively working in the current Haskell
ecosystem, we need to extend the versions of dependencies that Copilot
can be installed with. For example, optparse-applicative has seen a new
release 0.18.1.0, but copilot needs versions strictly lower than 0.18.

This commit relaxes the upper bounds on optparse-applicative so that it
can be installed with the 0.18 series.