copilot-theorem: Mark module as trustworthy. Refs #491.
The module Copilot.Theorem.Misc.Utils, which is marked as safe, imports
System.Directory, which in a recent minor update, has been marked as not
safe. This makes the import illegal.

This commit updates the module Copilot.Theorem.Misc.Utils as
trustworthy, so that the import of System.Directory does not lead to a
compilation error.