copilot-theorem: Add explicit imports for Control.Monad. Refs #491.
The modules in Control.Monad.* in mtl >= 2.3 no longer re-export
Control.Monad, leading to a number of compilation errors in
copilot-theorem. Since mtl >= 2.3 is required for Copilot to work with
GHC 9.6 and above, this issue prevents Copilot from compiling with
current versions of GHC, including the one in hackage.

This commit updates copilot-theorem so that any imports from
Control.Monad are explicit and not via re-exports from other modules.