copilot-theorem: Look for `falsifiable` (not `invalid`). Refs #495.
In Kind2, disproven properties are tagged with `falsifiable` in the XML output,
but the code in `copilot-theorem`'s Kind2 backend was instead searching for a
tag named `invalid`. As a result, `copilot-theorem` would error when attempting
to disprove properties that are false, as it fail to parse the XML output. This
fixes the issue by replacing `invalid` with `falsifiable`.