copilot-theorem: Remove mention of non-existent `Driver.hs` in README. Refs #452.
The section about combining provers mentioned this was shown in the
`Driver.hs` file, which does not exist anymore. Additionally, the 'Some
examples' section mentioned `Driver.hs` as part of the examples,
containing a main function. The `Driver.hs` file does not exist anymore,
so these references should be removed.