copilot-theorem: Update usage of `prove` in README. Refs #452.
The interface of `prove` has changed a lot. It no longer matches the
description. This commit updates all mentions of `prove`, both text and
examples, to match the current interface. The old examples used the
`kInduction` prover which no longer matches the interface. This commit
replaces those calls by the Kind2 prover.