copilot-c99: Expand expression translation environment to include statements. Refs #520.
Copilot currently supports reading values from fields of structs, but
not changing them. We'd like to be able to modify a struct by adjusting
the values of specific fields.

The current implementation of the translation of expressions generates
variable declarations, but we now need to keep a counter of the new
variables that have been created, and additional statements that must be
executed after the variables are declared.

This commit modifies the function environment to allow for such
elements, without adding any specific code that uses those. This change,
in isolation, should not modify the extensible behavior of Copilot.

Co-authored-by: Frank Dedden <dev@dedden.net>
Co-authored-by: Ryan Scott <rscott@galois.com>