copilot-c99: Handle new struct update operator. Refs #520.
Copilot currently supports reading values from fields of structs, but
not changing them. We'd like to be able to modify a struct by adjusting
the values of specific fields.

This commit extends the C99 backend to support updates to fields of
structs. We use a dedicated translation that injects statements as well
as new variable declarations, which are needed to calculate intermediate
values and the returned struct value in generators.

The original function that translated Copilot Op2's to C functions is
modified to not accept an UpdateField argument, as that case is now
handled separately due to its more complex functionality and type.

Co-authored-by: Frank Dedden <dev@dedden.net>
Co-authored-by: Ryan Scott <rscott@galois.com>