copilot-language: High-level array update operator. Refs #520.
Copilot currently supports reading values from fields of structs, but
not changing them. We'd like to be able to modify a struct by adjusting
the values of specific fields.

A prior commit has extended copilot-core with a low-level interface
to support struct field updates, and the C99 backend and the interpreter
have been extended accordingly.

This commit extends the high-level language definition to support struct
field updates. We need to add an auxiliary type Projection to provide
the notation we would like to support.

Co-authored-by: Frank Dedden <dev@dedden.net>
Co-authored-by: Ryan Scott <rscott@galois.com>