copilot-language: Remove deprecated function Copilot.Language.Spec.forall. Refs #518.
The function Copilot.Language.Spec.forall not being used by any other
part of Copilot. It was replaced by a function with a different name due
to the lowercase name not being allowed in newer versions of GHC. The
original function was deprecated in Copilot 3.18. As per our internal
policy of waiting 3 versions from deprecation until a public interface
declaration can be removed, this function can now be removed.

This commit removes the function Copilot.Language.Spec.forall from
copilot-language (both in the module that defines it and in a module
that re-exports it).