copilot-core: Add new update operation to Struct. Refs #520.
Copilot currently supports reading values from fields of structs, but
not changing them. We'd like to be able to modify a struct by adjusting
the values of specific fields.

In order to add support to the interpreter for Struct field updates, we
need an operation that will perform the modification in the specific
struct type itself (the one that is an instance of the Struct class).

This commit modifies the Struct class with the new operation. We use
Value as an argument, which imposes a Typeable and Show constraint,
which were added to the Op2 definition in a prior commit.

To make the class backwards compatible, we pre-define the method to an
error call that indicates to users that the operation has not been
defined for that specific type.

Co-authored-by: Frank Dedden <dev@dedden.net>
Co-authored-by: Ryan Scott <rscott@galois.com>