copilot-theorem: Support translating `UpdateField` to what4. Refs #524.
A new kind of `Op2`, called `UpdateField`, was added to `copilot-core` in a
prior commit to support defining streams with updates to struct fields.

This commit extends `copilot-theorem` to support reasoning about streams
involving struct updates. We do so by providing a new case for the operation
that translates a `copilot-core` `UpdateField` into a What4 representations of
a Copilot expression.

The implementation proves relatively straightforward, as struct field accessors
always correspond to a concrete index value, which can then be used to update
the value at the corresponding index in the list of struct fields.