copilot-core: Introduce operations to update arrays. Refs #36.
Copilot defines a types for fixed-length arrays, but no way to update
them. In order to implement array updates for values in streams, we need
to have a mechanism to update values in individual arrays.

This commit adds a function to arrays to update a value while preserving the
length.

Co-authored-by: Max Fan <maxwell.y.fan@nasa.gov>