copilot-theorem: Translate array updates to What4. Refs #36.
Copilot currently supports reading values from arrays, but not changing them.

This commit adds symbolic reasoning support for array update expressions by
translating them to What4. The implementation is largely inspired by the
existing translation for array index expressions. Because the index in an array
update expression can be symbolic, we must build a chain of symbolic
if-then-else expressions that check the value of the possibly-symbolic index
against all possible concrete index values from `0` to `n - 1` (where `n` is
the size of the array).