copilot-libraries: Rename function to avoid name clash. Refs #36.
We want to add support to be able to modify values of elements in
arrays. In order to provide a uniform interface for arrays and structs,
we have decided to unify the syntax for array and struct accessors, so
that one operator symbol indicates read access, and doubling the same
operator symbol indicates modification access. For example, (!) would
indicate accessing an element of an array to consult it, and (!!) would
indicate the same but to update it or modify it. The analogous operators
for structs would be (#) and (##).

At present, the symbol (!!) is being used for an auxiliary function that
indexes a list of streams by a stream of indexes. That function is
expensive, and rarely used in real specifications.

This commit renames such rarely used function, (!!), to avoid a name
clash with our desired name for the function to access elements of
arrays in streams for the purposes of modifying them.