copilot: Introduce example reasoning about array updates. Refs #36.
Previous commits have introduced support to modify individual elements
of arrays, either by giving a new stream of values for a position in an
array, or by applying a stream function to an element.

This commit introduces a new example in the examples/ directory to
demonstrate using copilot-theorem to reason about arrays where
individual elements are being updated.