copilot-language: Rename operator for uniformity. Refs #36.
We want to add support to be able to modify values of elements in
arrays. In order to provide a uniform interface for arrays and structs,
we have decided to unify the syntax for array and struct accessors, so
that one operator symbol indicates read access, and doubling the same
operator symbol indicates modification access. For example, (!) would
indicate accessing an element of an array to consult it, and (!!) would
indicate the same but to update it or modify it. The analogous operators
for structs would be (#) and (##).

At present, the symbol being used to access elements of arrays (.!!),
which is not consistent with the naming approach described above.

This commit defines a new operator (!) for array element access, and
deprecates the old (.!!).