copilot-language: Introduce new kind of array element update operation. Refs #36.
Copilot currently supports reading values from arrays, but not changing them.

This commit adds a new operation to the high-level Copilot language that
allows updating array elements. It relies on a low-level Op3 that
captures the correspnding operation in Copilot Core expressions.

In order to provide a uniform interface for arrays and structs, we have
decided to unify the syntax for array and struct accessors, so that the
operator symbol (=:) is used to set the value of individual elements,
and (=$) is used to apply a function to values of elements (at each
time). To support using the same syntax for both structs and arrays,
this commit moves those operators into a new class, defined in a module
Copilot.Language.Operators.Projection. The module
Copilot.Language.Operators.Struct now uses that module instead of
defining the operators locally, and so does the new module
Copilot.Language.Operators.Array.

Co-authored-by: Max Fan <maxwell.y.fan@nasa.gov>