copilot: Update what4-propositional example's comments to match results. Refs #535.
Previously, the comments in the `what4-propositional` example claimed that
certain examples were unprovable, even though they were actually provable. This
is a result of the heuristics in `Copilot.Theorem.What4` becoming more
intelligent after commit 4ae3f249fbf8c37c0909046150ff341d432bd042, but due to
an oversight, the comments in the `what4-propositional` example were not
updated at the time.

This commit revamps the `what4-propositional` example so that each example's
comments clearly and accurately state whether the example is provable with the
current implementation of `Copilot.Theorem.What4`. I have organized the
examples so that they are organized into logically separate sections, with each
section requiring increasingly more sophisticated techniques to prove (or, in
some cases, fail to prove).