copilot: Update contribution guidelines. Refs #476.
Copilot now has a process that allows us to accept contributions from
external users again.

This commit updates the README to describe the new process that users
should follow to contribute to Copilot. I am choosing to leave some
details out to not overwhelm users in the README; that can be discussed
later when the PR is processed.