copilot-core: Deprecate Copilot.Core.Type.UType.uTypeType. Refs #484.
The record field uTypeType, part of Copilot.Core.Type.UType, is never
really used within the same module or in any other part of Copilot or,
as far as we know, by any user of Copilot. Also, we cannot use it to
obtain the value inside the UType due to it being an existential type.
It also makes the test report lower coverage than they otherwise might
be, since the record access function is not and cannot be used by the
tests.

This field is not a necessary part of Copilot's interface, and not
required by Copilot's requirements. Therefore, it would be suitable to
remove it from the module.

This commit deprecates the field uTypeType, so that it can be removed in
a future version of Copilot.