copilot: Update struct examples to use generic method implementations. Refs #564.
Currently, using structs in Copilot requires defining instances of several type
classes. Defining those instances is unnecessarily cumbersome, and it exposes
the user to a substantial amount of Haskell even when they want to stay at the
level of the Copilot language.

In a previous commit, we added `Generic`-based defaults for the methods of the
`Struct` and `Typed` classes.

In this commit, we leverage these default implementations to remove much of the
boilerplate code needed to define `Struct` and `Typed` instances in the
struct-related `copilot` examples. Note that I have intentionally _not_ used
`updateFieldDefault` in the `StructsUpdateField.hs` example, as that example is
intended to demonstrate how one would implement `updateField` by hand.

This work is based off of an initial implementation by Marten Wijnja (@Qqwy).