copilot-core: Warn about future generic default for updateField. Refs #564.
Currently, using structs in Copilot requires defining instances of several type
classes. Defining those instances is unnecessarily cumbersome, and it exposes
the user to a substantial amount of Haskell even when they want to stay at the
level of the Copilot language.

In the previous commit, we added `Generic`-based defaults for several type
class methods, including the `updateField` method of `Struct`.

In the future, we plan to change the default implementation for `updateField`
such that it will require a `Generic` instance. This will break any existing
`Struct` instance that omits an implementation of `updateField` and also does
not define a `Generic` instance for the struct data type. Unfortunately, there
does not appear to be a way for GHC to warn about this combination of
properties.

This commit adds a verbal warning to the `updateField` Haddocks to make users
aware of this potential problem.