copilot-core: Warn about future generic default for updateField. Refs #564.
In the future, we plan to change the default implementation for `updateField`
such that it will require a `Generic` instance. This will break any existing
`Struct` instance that omits an implementation of `updateField` and also does
not define a `Generic` instance for the struct data type. Unfortunately, there
does not appear to be a way for GHC to warn about this combination of
properties, but we can at least warn about this in the `updateField` Haddocks.