copilot-core: Deprecate Copilot.Core.Type.UType.uExprType. Refs #565.
The record fields uExprType and uExprExpr, part of Copilot.Core.Expr.UExpr, are
never really used within the module and only rarely used elsewhere in Copilot
or, as far as we know, by any user of Copilot. As a matter of fact, we cannot
use them to obtain the value inside the UExpr due to the existential it
encloses.

These fields are not necessary parts of the interface to use Copilot, and not
required by Copilot's requirements. Therefore, it would be suitable to remove
that definitions from the module.

This commit removes any uses of uExprType and uExprExpr incopilot-theorem and
uses positional arguments instead.