copilot-c99: Add specific `Trigger` representation. Refs #296.
The current implementation of Copilot does not allow using the same
trigger handler name multiple times, since the C99 backend produces
multiple repeated declarations, which is invalid. Consequently, the code
generator assumes that trigger names are unique.

To allow for multiple triggers that trigger the same handler, we need to
introduce a separate construct in triggers that acts as a unique
identifier.

This commit introduces an internal datatype to represent a `Trigger`
with a unique name that is different from that of the function called
when triggered.